[refactor] 본인인증 입력 폼 공통 컴포넌트 분리 #275
